Merge to master for release

+++ AppCenterCrashes/AppCenterCrashes/Internals/MSCrashesCXXExceptionHandler.mm
@@ -6,7 +6,6 @@
 #import <pthread.h>
 #import <stdexcept>
 #import <string>
-#import <typeinfo>
 #import <vector>
 
 #import "MSCrashesCXXExceptionHandler.h"
@@ -30,21 +29,16 @@ static pthread_key_t _MSCrashesCXXExceptionInfoTSDKey = 0;
 
 @implementation MSCrashesUncaughtCXXExceptionHandlerManager
 
-extern "C" void __attribute__((noreturn))
-__cxa_throw(void *exception_object, std::type_info *tinfo, void (*dest)(void *)) {
-  
+extern "C" void __attribute__((noreturn)) __cxa_throw(void *exception_object, std::type_info *tinfo, void (*dest)(void *)) {
+
   /*
-   * Purposely do not take a lock in this function. The aim is to be as fast as
-   * possible. While we could really use some of the info set up by the real
-   * __cxa_throw, if we call through we never get control back - the function is
-   * noreturn and jumps to landing pads. Most of the stuff in __cxxabiv1 also
-   * won't work yet. We therefore have to do these checks by hand.
+   * Purposely do not take a lock in this function. The aim is to be as fast as possible. While we could really use some of the info set up
+   * by the real __cxa_throw, if we call through we never get control back - the function is noreturn and jumps to landing pads. Most of the
+   * stuff in __cxxabiv1 also won't work yet. We therefore have to do these checks by hand.
    *
-   * The technique for distinguishing Objective-C exceptions is based on the
-   * implementation of objc_exception_throw(). It's weird, but it's fast. The
-   * explicit symbol load and NULL checks should guard against the
-   * implementation changing in a future version. (Or not existing in an earlier
-   * version).
+   * The technique for distinguishing Objective-C exceptions is based on the implementation of objc_exception_throw(). It's weird, but it's
+   * fast. The explicit symbol load and NULL checks should guard against the implementation changing in a future version. (Or not existing
+   * in an earlier version).
    */
   typedef void (*cxa_throw_func)(void *, std::type_info *, void (*)(void *)) __attribute__((noreturn));
   static dispatch_once_t predicate = 0;
@@ -56,24 +50,19 @@ __cxa_throw(void *exception_object, std::type_info *tinfo, void (*dest)(void *))
     __real_objc_ehtype_vtable = reinterpret_cast<const void **>(dlsym(RTLD_DEFAULT, "objc_ehtype_vtable"));
   });
 
-  /*
-   *   Actually check for Objective-C exceptions.
-   */
+  // Actually check for Objective-C exceptions.
   if (tinfo && __real_objc_ehtype_vtable && // Guard from an ABI change
       *reinterpret_cast<void **>(tinfo) == __real_objc_ehtype_vtable + 2) {
     goto callthrough;
   }
 
   /*
-   * Any other exception that came here has to be C++, since Objective-C is the
-   * only (known) runtime that hijacks the C++ ABI this way. We need to save off
-   * a backtrace.
-   * Invariant: If the terminate handler is installed, the TSD key must also be
-   * initialized.
+   * Any other exception that came here has to be C++, since Objective-C is the only (known) runtime that hijacks the C++ ABI this way. We
+   * need to save off a backtrace.
+   * Invariant: If the terminate handler is installed, the TSD key must also be initialized.
    */
   if (_MSCrashesIsOurTerminateHandlerInstalled) {
-    MSCrashesCXXExceptionTSInfo *info =
-        static_cast<MSCrashesCXXExceptionTSInfo *>(pthread_getspecific(_MSCrashesCXXExceptionInfoTSDKey));
+    MSCrashesCXXExceptionTSInfo *info = static_cast<MSCrashesCXXExceptionTSInfo *>(pthread_getspecific(_MSCrashesCXXExceptionInfoTSDKey));
 
     if (!info) {
       info = reinterpret_cast<MSCrashesCXXExceptionTSInfo *>(calloc(1, sizeof(MSCrashesCXXExceptionTSInfo)));
@@ -81,8 +70,8 @@ __cxa_throw(void *exception_object, std::type_info *tinfo, void (*dest)(void *))
     }
     info->exception_object = exception_object;
     // XXX: All significant time in this call is spent right here.
-    info->num_frames = backtrace(reinterpret_cast<void **>(&info->call_stack[0]),
-                                 sizeof(info->call_stack) / sizeof(info->call_stack[0]));
+    info->num_frames = static_cast<uint32_t>(
+        backtrace(reinterpret_cast<void **>(&info->call_stack[0]), sizeof(info->call_stack) / sizeof(info->call_stack[0])));
   }
 
 callthrough:
@@ -127,36 +116,33 @@ static void MSCrashesUncaughtCXXTerminateHandler(void) {
         info.exception_frames_count = recorded_info->num_frames - 1;
         info.exception_frames = &recorded_info->call_stack[1];
       } else {
-        
-        /*
-         * There's no backtrace, grab this function's trace instead. Probably
-         * means the exception came from a dynamically loaded library.
-         */
+
+        // There's no backtrace, grab this function's trace instead. Probably means the exception came from a dynamically loaded library.
         void *frames[128] = {nullptr};
 
-        info.exception_frames_count = backtrace(&frames[0], sizeof(frames) / sizeof(frames[0])) - 1;
+        info.exception_frames_count = static_cast<uint32_t>(backtrace(&frames[0], sizeof(frames) / sizeof(frames[0])) - 1);
         info.exception_frames = reinterpret_cast<uintptr_t *>(&frames[1]);
       }
 
       try {
         std::rethrow_exception(p);
       } catch (const std::exception &e) {
-        
+
         // C++ exception.
         info.exception_message = e.what();
         MSCrashesIterateExceptionHandlers_unlocked(info);
       } catch (const std::exception *e) {
-        
+
         // C++ exception by pointer.
         info.exception_message = e->what();
         MSCrashesIterateExceptionHandlers_unlocked(info);
       } catch (const std::string &e) {
-        
+
         // C++ string as exception.
         info.exception_message = e.c_str();
         MSCrashesIterateExceptionHandlers_unlocked(info);
       } catch (const std::string *e) {
-        
+
         // C++ string pointer as exception.
         info.exception_message = e->c_str();
         MSCrashesIterateExceptionHandlers_unlocked(info);
@@ -164,7 +150,7 @@ static void MSCrashesUncaughtCXXTerminateHandler(void) {
         info.exception_message = e;
         MSCrashesIterateExceptionHandlers_unlocked(info);
       } catch (__attribute__((unused)) id e) {
-        
+
         // Objective-C exception. Pass it on to Foundation.
         OSSpinLockUnlock(&_MSCrashesCXXExceptionHandlingLock);
         if (_MSCrashesOriginalTerminateHandler != nullptr) {
@@ -172,14 +158,14 @@ static void MSCrashesUncaughtCXXTerminateHandler(void) {
         }
         return;
       } catch (...) {
-        
+
         // Any other kind of exception. No message.
         MSCrashesIterateExceptionHandlers_unlocked(info);
       }
     }
   }
   OSSpinLockUnlock(&_MSCrashesCXXExceptionHandlingLock);
-  
+
   // In case terminate is called reentrantly by passing it on.
   if (_MSCrashesOriginalTerminateHandler != nullptr) {
     _MSCrashesOriginalTerminateHandler();
@@ -191,10 +177,7 @@ static void MSCrashesUncaughtCXXTerminateHandler(void) {
 + (void)addCXXExceptionHandler:(MSCrashesUncaughtCXXExceptionHandler)handler {
   static dispatch_once_t key_predicate = 0;
 
-  /*
-   * This only EVER has to be done once, since we don't delete the TSD later
-   * (there's no reason to delete it).
-   */
+  // This only EVER has to be done once, since we don't delete the TSD later (there's no reason to delete it).
   dispatch_once(&key_predicate, ^{
     pthread_key_create(&_MSCrashesCXXExceptionInfoTSDKey, free);
   });
@@ -213,8 +196,7 @@ static void MSCrashesUncaughtCXXTerminateHandler(void) {
 + (void)removeCXXExceptionHandler:(MSCrashesUncaughtCXXExceptionHandler)handler {
   OSSpinLockLock(&_MSCrashesCXXExceptionHandlingLock);
   {
-    auto i = std::find(_MSCrashesUncaughtExceptionHandlerList.begin(), _MSCrashesUncaughtExceptionHandlerList.end(),
-                       handler);
+    auto i = std::find(_MSCrashesUncaughtExceptionHandlerList.begin(), _MSCrashesUncaughtExceptionHandlerList.end(), handler);
 
     if (i != _MSCrashesUncaughtExceptionHandlerList.end()) {
       _MSCrashesUncaughtExceptionHandlerList.erase(i);
